+++ konte/index/__init__.py
@@ -2,6 +2,7 @@
 
 from konte.index.base import Store
 from konte.index.bm25_store import BM25Store
+from konte.index.embeddings import clear_query_embedding_cache, query_vector
 from konte.index.faiss_store import FAISSStore
 from konte.index.filter_index import FilterIndex, matches_filter_value
 
@@ -10,5 +11,7 @@ __all__ = [
     "FAISSStore",
     "FilterIndex",
     "Store",
+    "clear_query_embedding_cache",
     "matches_filter_value",
+    "query_vector",
 ]